requirements.txt
altgraph==0.17.4
bottle==0.13.2
cffi==1.17.1
clr_loader==0.2.7.post0
greenlet==3.1.1
packaging==24.2
pefile==2023.2.7
pip==24.3.1
proxy_tools==0.1.0
pycparser==2.22
pyinstaller==6.11.1
pyinstaller-hooks-contrib==2024.10
pythonnet==3.0.5
pywebview==5.3.2
pywin32-ctypes==0.2.3
setuptools==75.6.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2